implement refresh timer and file sd generation
This implements a method for utilizing the `X-Prometheus-Refresh-Interval` header.
Currently the implementation aims to Update as often as the fastest Prometheus
polling this. There is no way to increase the timer (except for a restart) at the
moment.

I've also implemented a oneshot run mode, that output a Prometheus file SD.

Also ported the application to a kong application.

Signed-off-by: Francis Begyn <francis@begyn.be>